+++ bin/commands/on.rb
@@ -13,20 +13,6 @@ command :on do |c|
   c.arg_name 'NAME'
   c.flag %i[s section], default_value: 'All'
 
-  c.desc 'Show time intervals on @done tasks'
-  c.switch %i[t times], default_value: true, negatable: true
-
-  c.desc 'Show elapsed time on entries without @done tag'
-  c.switch [:duration]
-
-  c.desc 'Show time totals at the end of output'
-  c.switch [:totals], default_value: false, negatable: false
-
-  c.desc 'Sort tags by (name|time)'
-  default = Doing.setting('tag_sort').normalize_tag_sort || :name
-  c.arg_name 'KEY'
-  c.flag [:tag_sort], must_match: REGEX_TAG_SORT, default_value: default, type: TagSortSymbol
-
   c.desc "Output to export format (#{Doing::Plugins.plugin_names(type: :export)})"
   c.arg_name 'FORMAT'
   c.flag %i[o output]
@@ -38,6 +24,11 @@ command :on do |c|
   c.desc 'Override output format with a template string containing %placeholders'
   c.arg_name 'TEMPLATE_STRING'
   c.flag [:template]
+
+  add_options(:time_display, c)
+  add_options(:search, c)
+  add_options(:tag_filter, c)
+  add_options(:time_filter, c)
 
   c.action do |_global_options, options, args|
     raise DoingRuntimeError, %(Invalid output type "#{options[:output]}") if options[:output] && options[:output] !~ Doing::Plugins.plugin_regex(type: :export)
@@ -59,7 +50,6 @@ command :on do |c|
     options[:times] = true if options[:totals]
     options[:sort_tags] = options[:tag_sort]
 
-    Doing::Pager.page @wwid.list_date([start, finish], options[:section], options[:times], options[:output],
-                        { template: options[:template], config_template: options[:config_template], duration: options[:duration], totals: options[:totals], sort_tags: options[:sort_tags] }).chomp
+    Doing::Pager.page @wwid.list_date([start, finish], options[:section], options[:times], options[:output], options).chomp
   end
 end