+++ lib/doing/items.rb
@@ -58,6 +58,24 @@ module Doing
       Doing.logger.info('New section:', %("#{section}" added)) if log
     end
 
+    def delete_section(section, log: false)
+      return unless section?(section)
+
+      raise DoingRuntimeError, 'Section not empty' if in_section(section).count > 0
+
+      deleted = false
+
+      @sections.each do |sect|
+        if sect.title == section && in_section(sect).count.zero?
+          @sections.delete(sect)
+          Doing.logger.info('Removed section:', %("#{section}" removed)) if log
+          return
+        end
+      end
+
+      Doing.logger.error('Not found:', %("#{section}" not found))
+    end
+
     # Get a new Items object containing only items in a
     # specified section
     #
@@ -124,6 +142,36 @@ module Doing
         diff.push(item) unless res.count.positive?
       end
       diff
+    end
+
+    ##
+    ## Remove duplicated entries. Duplicate entries must have matching start date, title, note, and section
+    ##
+    ## @return     [Items] Items array with duplicate entries removed
+    ##
+    def dedup(match_section: true)
+      unique = Items.new
+      each do |item|
+        unique.push(item) unless unique.include?(item, match_section: match_section)
+      end
+
+      unique
+    end
+
+    def dedup!(match_section: true)
+      replace dedup(match_section: match_section)
+    end
+
+    def include?(item, match_section: true)
+      includes = false
+      each do |other_item|
+        if other_item.equal?(item, match_section: match_section)
+          includes = true
+          break
+        end
+      end
+
+      includes
     end
 
     # Output sections and items in Doing file format