+++ lib/doing/types.rb
@@ -12,7 +12,7 @@ module Doing
     REGEX_DAY = /^(mon|tue|wed|thur?|fri|sat|sun)(\w+(day)?)?$/i.freeze
     REGEX_RANGE_INDICATOR = ' +(?:to|through|thru|(?:un)?til|-+) +'
     REGEX_RANGE = /^\S+.*?#{REGEX_RANGE_INDICATOR}\S+.*?$/i.freeze
-    REGEX_TIME_RANGE = /^#{REGEX_CLOCK}#{REGEX_RANGE_INDICATOR}#{REGEX_CLOCK}$/i.freeze
+    REGEX_TIME_RANGE = /^#{REGEX_CLOCK}(?:#{REGEX_RANGE_INDICATOR}#{REGEX_CLOCK})?$/i.freeze
 
     InvalidExportType = Class.new(RuntimeError)
     MissingConfigFile = Class.new(RuntimeError)